Separated tests out into separate file.
